+++ changelog.md
@@ -1,3 +1,2 @@
-* Añadido nueva carga de la interface al actualizar la base de datos.
-* Preparación del complemento para la compatibilidad con NVDA 2024.1.
-* Añadida nueva búsqueda inmediata por @javidominguez
+* Añadido guardar posición de reproducción.
+* Solución de errores.